3-layer arhitektura
