+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/DTOs/Notifications/NotificationDto.cs
@@ -1,4 +1,4 @@
-using TelecomCustomerSupportSystem.Domain.Enums;
+using TelecomCustomerSupportSystem.Application.Enums;
 
 namespace TelecomCustomerSupportSystem.Application.DTOs.Notifications;
