+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/Entities/Komentar.cs
@@ -1,4 +1,4 @@
-namespace TelecomCustomerSupportSystem.Domain.Entities;
+namespace TelecomCustomerSupportSystem.Application.Entities;
 
 public class Komentar
 {
