+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/Enums/Prioritet.cs
@@ -1,4 +1,4 @@
-namespace TelecomCustomerSupportSystem.Domain.Enums;
+namespace TelecomCustomerSupportSystem.Application.Enums;
 
 public enum Prioritet
 {
