+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/Enums/StatusPaketa.cs
@@ -1,4 +1,4 @@
-namespace TelecomCustomerSupportSystem.Domain.Enums;
+namespace TelecomCustomerSupportSystem.Application.Enums;
 
 public enum StatusPaketa
 {
