+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/Enums/StatusTiketa.cs
@@ -1,4 +1,4 @@
-namespace TelecomCustomerSupportSystem.Domain.Enums;
+namespace TelecomCustomerSupportSystem.Application.Enums;
 
 public enum StatusTiketa
 {
