+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/Enums/TipTima.cs
@@ -1,0 +1,7 @@
+namespace TelecomCustomerSupportSystem.Application.Enums;
+
+public enum TipTima
+{
+    AGENTI = 1,
+    TEHNICARI = 2
+}
