+++ Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Application/TelecomCustomerSupportSystem.Application.csproj
@@ -1,11 +1,7 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\TelecomCustomerSupportSystem.Domain\TelecomCustomerSupportSystem.Domain.csproj" />
-  </ItemGroup>
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
