--- Project/TelecomCustomerSupportSystem/backend/TelecomCustomerSupportSystem.Domain/Enums/TipTima.cs
@@ -1,7 +1,0 @@
-namespace TelecomCustomerSupportSystem.Domain.Enums;
-
-public enum TipTima
-{
-    AGENTI = 1,
-    TEHNICARI = 2
-}
